Move hosting view to hCore
